@@ -5,11 +5,11 @@ plugins {
 
 android {
     namespace 'com.livinglifetechway.quickpermissions_sample'
-    compileSdkVersion 27
+    compileSdk 33
     defaultConfig {
         applicationId "com.livinglifetechway.quickpermissions_sample"
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdk 33
         versionCode 1
         versionName "1.0"
     }
@@ -23,11 +23,16 @@ android {
 }
 
 dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.6.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.github.kirtan403.k4kotlin:k4kotlin:0.5.4'
+    implementation 'com.google.android.material:material:1.8.0'
     implementation 'org.jetbrains.anko:anko-commons:0.10.5'
     implementation project(':quickpermissions-kotlin')
+}
+
+configurations {
+    all {
+        exclude group: 'com.android.support'
+    }
 }
