@@ -2,7 +2,7 @@ package com.livinglifetechway.quickpermissions_sample
 
 import android.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.livinglifetechway.quickpermissions_sample.databinding.ActivityMainBinding
 import com.livinglifetechway.quickpermissions_sample.kotlin.AllKotlinActivity
 import org.jetbrains.anko.startActivity
