@@ -3,8 +3,8 @@ package com.livinglifetechway.quickpermissions_kotlin.util
 import android.app.Activity
 import android.content.Context
 import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
-import android.support.v4.app.Fragment
+import androidx.core.app.ActivityCompat
+import androidx.fragment.app.Fragment
 
 /**
  * Utility class that wraps access to the runtime permissions API in M and provides basic helper
