@@ -1,9 +1,9 @@
 package com.livinglifetechway.quickpermissions_kotlin
 
 import android.content.Context
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
 import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
 import com.livinglifetechway.quickpermissions_kotlin.util.PermissionCheckerFragment
 import com.livinglifetechway.quickpermissions_kotlin.util.PermissionsUtil
 import com.livinglifetechway.quickpermissions_kotlin.util.QuickPermissionsRequest
