@@ -6,9 +6,9 @@ import android.content.pm.PackageManager
 import android.net.Uri.fromParts
 import android.os.Bundle
 import android.provider.Settings.ACTION_APPLICATION_DETAILS_SETTINGS
-import android.support.v4.app.ActivityCompat
-import android.support.v4.app.Fragment
 import android.util.Log
+import androidx.core.app.ActivityCompat
+import androidx.fragment.app.Fragment
 import org.jetbrains.anko.alert
 
 /**
